+++ data_class_annotation_generator/CHANGELOG.md
@@ -22,6 +22,8 @@ Collection equality
 
 Custom equality
 
-## 0.1.7
+## 0.1.8
 
-.
+Bump dependencies versions
+Added parameters to chose which method to generate
+Fix issue with extended classes